+++ docs/source/conf.py
@@ -10,7 +10,9 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
+import glob
 import os
+import shutil
 import sys
 from importlib.util import module_from_spec, spec_from_file_location
 
@@ -22,6 +24,7 @@ sys.path.insert(0, os.path.abspath(_PATH_ROOT))
 
 try:
     from flash import __about__ as about
+    from flash.core.utilities import providers
 
 except ModuleNotFoundError:
 
@@ -32,6 +35,7 @@ except ModuleNotFoundError:
         return py
 
     about = _load_py_module("__about__.py")
+    providers = _load_py_module("flash/core/utilities/providers.py")
 
 SPHINX_MOCK_REQUIREMENTS = int(os.environ.get("SPHINX_MOCK_REQUIREMENTS", True))
 
@@ -42,6 +46,45 @@ html_favicon = "_static/images/icon.svg"
 project = "Flash"
 copyright = "2020-2021, PyTorch Lightning"
 author = "PyTorch Lightning"
+
+# -- Project documents -------------------------------------------------------
+
+
+def _transform_changelog(path_in: str, path_out: str) -> None:
+    with open(path_in) as fp:
+        chlog_lines = fp.readlines()
+    # enrich short subsub-titles to be unique
+    chlog_ver = ""
+    for i, ln in enumerate(chlog_lines):
+        if ln.startswith("## "):
+            chlog_ver = ln[2:].split("-")[0].strip()
+        elif ln.startswith("### "):
+            ln = ln.replace("###", f"### {chlog_ver} -")
+            chlog_lines[i] = ln
+    with open(path_out, "w") as fp:
+        fp.writelines(chlog_lines)
+
+
+generated_dir = os.path.join(_PATH_HERE, "generated")
+
+os.makedirs(generated_dir, exist_ok=True)
+# copy all documents from GH templates like contribution guide
+for md in glob.glob(os.path.join(_PATH_ROOT, ".github", "*.md")):
+    shutil.copy(md, os.path.join(generated_dir, os.path.basename(md)))
+# copy also the changelog
+_transform_changelog(os.path.join(_PATH_ROOT, "CHANGELOG.md"), os.path.join(generated_dir, "CHANGELOG.md"))
+
+# -- Generate providers ------------------------------------------------------
+
+lines = []
+for provider in providers.PROVIDERS:
+    lines.append(f"- {str(provider)}\n")
+
+generated_dir = os.path.join("integrations", "generated")
+os.makedirs(generated_dir, exist_ok=True)
+
+with open(os.path.join(generated_dir, "providers.rst"), "w") as f:
+    f.writelines(sorted(lines, key=str.casefold))
 
 # -- General configuration ---------------------------------------------------
 
@@ -79,7 +122,7 @@ templates_path = ["_templates"]
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = []
+exclude_patterns = ["generated/PULL_REQUEST_TEMPLATE.md"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
@@ -140,7 +183,7 @@ def setup(app):
 # https://stackoverflow.com/questions/15889621/sphinx-how-to-exclude-imports-in-automodule
 def _package_list_from_file(pfile):
     assert os.path.isfile(pfile)
-    with open(pfile, "r") as fp:
+    with open(pfile) as fp:
         lines = fp.readlines()
     list_pkgs = []
     for ln in lines:
@@ -162,7 +205,7 @@ PACKAGE_MAPPING = {
     "pytorch-tabnet": "pytorch_tabnet",
     "pyDeprecate": "deprecate",
 }
-MOCK_PACKAGES = []
+MOCK_PACKAGES = ["numpy", "PyYAML", "tqdm"]
 if SPHINX_MOCK_REQUIREMENTS:
     # mock also base packages when we are on RTD since we don't install them there
     MOCK_PACKAGES += _package_list_from_file(os.path.join(_PATH_ROOT, "requirements.txt"))