+++ flash/image/classification/backbones/resnet.py
@@ -62,7 +62,7 @@ class BasicBlock(nn.Module):
         dilation: int = 1,
         norm_layer: Optional[Callable[..., nn.Module]] = None,
     ) -> None:
-        super(BasicBlock, self).__init__()
+        super().__init__()
         if norm_layer is None:
             norm_layer = nn.BatchNorm2d
         if groups != 1 or base_width != 64:
@@ -118,7 +118,7 @@ class Bottleneck(nn.Module):
         dilation: int = 1,
         norm_layer: Optional[Callable[..., nn.Module]] = None,
     ) -> None:
-        super(Bottleneck, self).__init__()
+        super().__init__()
         if norm_layer is None:
             norm_layer = nn.BatchNorm2d
         width = int(planes * (base_width / 64.0)) * groups
@@ -169,9 +169,10 @@ class ResNet(nn.Module):
         norm_layer: Optional[Callable[..., nn.Module]] = None,
         first_conv3x3: bool = False,
         remove_first_maxpool: bool = False,
+        in_chans: int = 3,
     ) -> None:
 
-        super(ResNet, self).__init__()
+        super().__init__()
 
         if norm_layer is None:
             norm_layer = nn.BatchNorm2d
@@ -194,9 +195,9 @@ class ResNet(nn.Module):
         num_out_filters = width_per_group * widen
 
         if first_conv3x3:
-            self.conv1 = nn.Conv2d(3, num_out_filters, kernel_size=3, stride=1, padding=1, bias=False)
+            self.conv1 = nn.Conv2d(in_chans, num_out_filters, kernel_size=3, stride=1, padding=1, bias=False)
         else:
-            self.conv1 = nn.Conv2d(3, num_out_filters, kernel_size=7, stride=2, padding=3, bias=False)
+            self.conv1 = nn.Conv2d(in_chans, num_out_filters, kernel_size=7, stride=2, padding=3, bias=False)
 
         self.bn1 = norm_layer(num_out_filters)
         self.relu = nn.ReLU(inplace=True)
@@ -320,7 +321,7 @@ def _resnet(
     model_weights = None
     if pretrained_flag:
         if "supervised" not in weights_paths:
-            raise KeyError("Supervised pretrained weights not available for {0}".format(model_name))
+            raise KeyError(f"Supervised pretrained weights not available for {model_name}")
 
         model_weights = load_state_dict_from_url(
             weights_paths["supervised"], map_location=torch.device("cpu") if device == -1 else torch.device(device)