+++ flash/pointcloud/detection/model.py
@@ -163,8 +163,7 @@ class PointCloudObjectDetector(Task):
         shuffle: bool = False,
         drop_last: bool = True,
         sampler: Optional[Sampler] = None,
-        convert_to_dataloader: bool = True,
-    ) -> Union[DataLoader, BaseAutoDataset]:
+    ) -> DataLoader:
 
         if not _POINTCLOUD_AVAILABLE:
             raise ModuleNotFoundError("Please, run `pip install flash[pointcloud]`.")
@@ -172,17 +171,13 @@ class PointCloudObjectDetector(Task):
         dataset.preprocess_fn = self.model.preprocess
         dataset.transform_fn = self.model.transform
 
-        if convert_to_dataloader:
-            return DataLoader(
-                dataset,
-                batch_size=batch_size,
-                num_workers=num_workers,
-                pin_memory=pin_memory,
-                collate_fn=collate_fn,
-                shuffle=shuffle,
-                drop_last=drop_last,
-                sampler=sampler,
-            )
-
-        else:
-            return dataset
+        return DataLoader(
+            dataset,
+            batch_size=batch_size,
+            num_workers=num_workers,
+            pin_memory=pin_memory,
+            collate_fn=collate_fn,
+            shuffle=shuffle,
+            drop_last=drop_last,
+            sampler=sampler,
+        )