+++ tests/core/test_model.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import math
+from itertools import chain
 from numbers import Number
 from pathlib import Path
 from typing import Any, Tuple
@@ -28,9 +29,10 @@ from torch.nn import functional as F
 from torch.utils.data import DataLoader
 
 import flash
+from flash.core.adapter import Adapter
 from flash.core.classification import ClassificationTask
 from flash.core.data.process import DefaultPreprocess, Postprocess
-from flash.core.utilities.imports import _PIL_AVAILABLE, _TABULAR_AVAILABLE, _TEXT_AVAILABLE
+from flash.core.utilities.imports import _TABULAR_AVAILABLE, _TEXT_AVAILABLE, Image
 from flash.image import ImageClassificationData, ImageClassifier
 from tests.helpers.utils import _IMAGE_TESTING, _TABULAR_TESTING
 
@@ -39,26 +41,25 @@ if _TABULAR_AVAILABLE:
 else:
     TabularClassifier = None
 
-if _PIL_AVAILABLE:
-    from PIL import Image
-else:
-
-    class Image:
-        Image = None
-
 
 # ======== Mock functions ========
 
 
 class DummyDataset(torch.utils.data.Dataset):
+    def __init__(self, num_samples: int = 9):
+        self.num_samples = num_samples
+
     def __getitem__(self, index: int) -> Tuple[Tensor, Number]:
         return torch.rand(1, 28, 28), torch.randint(10, size=(1,)).item()
 
     def __len__(self) -> int:
-        return 9
+        return self.num_samples
 
 
 class PredictDummyDataset(DummyDataset):
+    def __init__(self, num_samples: int):
+        super().__init__(num_samples)
+
     def __getitem__(self, index: int) -> Tensor:
         return torch.rand(1, 28, 28)
 
@@ -118,6 +119,30 @@ class GrandParent(Parent):
         super().__init__(Parent(child))
 
 
+class BasicAdapter(Adapter):
+    def __init__(self, child):
+        super().__init__()
+
+        self.child = child
+
+    def training_step(self, batch, batch_idx):
+        return self.child.training_step(batch, batch_idx)
+
+    def validation_step(self, batch, batch_idx):
+        return self.child.validation_step(batch, batch_idx)
+
+    def test_step(self, batch, batch_idx):
+        return self.child.test_step(batch, batch_idx)
+
+    def forward(self, x):
+        return self.child(x)
+
+
+class AdapterParent(Parent):
+    def __init__(self, child):
+        super().__init__(BasicAdapter(child))
+
+
 # ================================
 
 
@@ -133,7 +158,7 @@ def test_classificationtask_train(tmpdir: str, metrics: Any):
     assert "test_nll_loss" in result[0]
 
 
-@pytest.mark.parametrize("task", [Parent, GrandParent])
+@pytest.mark.parametrize("task", [Parent, GrandParent, AdapterParent])
 def test_nested_tasks(tmpdir, task):
     model = nn.Sequential(nn.Flatten(), nn.Linear(28 * 28, 10), nn.Softmax())
     train_dl = torch.utils.data.DataLoader(DummyDataset())
@@ -186,15 +211,12 @@ def test_classification_task_predict_folder_path(tmpdir):
 def test_classification_task_trainer_predict(tmpdir):
     model = nn.Sequential(nn.Flatten(), nn.Linear(28 * 28, 10))
     task = ClassificationTask(model)
-    ds = PredictDummyDataset()
-    batch_size = 3
-    predict_dl = torch.utils.data.DataLoader(ds, batch_size=batch_size)
+    ds = PredictDummyDataset(10)
+    batch_size = 6
+    predict_dl = task.process_predict_dataset(ds, batch_size=batch_size)
     trainer = pl.Trainer(default_root_dir=tmpdir)
     predictions = trainer.predict(task, predict_dl)
-    assert len(predictions) == len(ds) // batch_size
-    for batch_pred in predictions:
-        assert len(batch_pred) == batch_size
-        assert all(y < 10 for y in batch_pred)
+    assert len(list(chain.from_iterable(predictions))) == 10
 
 
 def test_task_datapipeline_save(tmpdir):
@@ -259,7 +281,7 @@ def test_available_backbones():
     class Foo(ImageClassifier):
         backbones = None
 
-    assert Foo.available_backbones() == []
+    assert Foo.available_backbones() == {}
 
 
 def test_optimization(tmpdir):
@@ -309,7 +331,7 @@ def test_optimization(tmpdir):
             scheduler_kwargs={"num_warmup_steps": 0.1},
             loss_fn=F.nll_loss,
         )
-        trainer = flash.Trainer(max_epochs=1, limit_train_batches=2)
+        trainer = flash.Trainer(max_epochs=1, limit_train_batches=2, gpus=torch.cuda.device_count())
         ds = DummyDataset()
         trainer.fit(task, train_dataloader=DataLoader(ds))
         optimizer, scheduler = task.configure_optimizers()
@@ -330,5 +352,5 @@ def test_classification_task_metrics():
             assert math.isclose(trainer.callback_metrics["train_accuracy_epoch"], 0.5)
 
     task = ClassificationTask(model)
-    trainer = flash.Trainer(max_epochs=1, callbacks=CheckAccuracy())
+    trainer = flash.Trainer(max_epochs=1, callbacks=CheckAccuracy(), gpus=torch.cuda.device_count())
     trainer.fit(task, train_dataloader=DataLoader(train_dataset), val_dataloaders=DataLoader(val_dataset))